remove ArrayPacker and unused ArrayPacker utilities from everything but bptt training
