+++ external/fv3fit/fv3fit/__init__.py
@@ -1,4 +1,4 @@
-from ._shared import ArrayPacker, StandardScaler, DerivedModel
+from ._shared import StandardScaler, DerivedModel
 from ._shared.input_sensitivity import (
     RandomForestInputSensitivities,
     JacobianInputSensitivity,
