+++ external/fv3fit/fv3fit/keras/__init__.py
@@ -6,5 +6,4 @@ from ._models.shared.training_loop import (
     TrainingLoopConfig,
     EpochResult,
 )
-from ._models.shared.utils import get_input_vector
 from ._models.recurrent import StepwiseModel
