+++ external/fv3fit/fv3fit/keras/_models/recurrent.py
@@ -1,18 +1,22 @@
-from typing import Iterable, Optional, Sequence, Tuple, Hashable
+import dataclasses
+from typing import Dict, Iterable, List, Optional, Sequence, TextIO, Tuple, Hashable
+from fv3fit._shared.config import PackerConfig
+from fv3fit._shared.packer import count_features, pack, unpack
+from fv3fit.keras._models.packer import Unpack
 import tensorflow as tf
 import xarray as xr
 from ..._shared import (
     io,
-    ArrayPacker,
     SAMPLE_DIM_NAME,
 )
-from .packer import get_unpack_layer
+import dacite
+import pandas as pd
 from .normalizer import LayerStandardScaler
 import numpy as np
 import vcm.safe
 import os
 import yaml
-from fv3fit.keras._models.shared import get_input_vector, PureKerasModel
+from fv3fit.keras._models.shared import PureKerasModel
 from ..._shared import get_dir, put_dir
 
 Z_DIMS = ["z", "z_interface"]
@@ -54,6 +58,164 @@ def _moisture_tendency_limiter(packer, state_update, state):
 
     dQ2 = tf.keras.layers.Lambda(limit)([dQ2, Q2])
     return pack([dQ1, dQ2])
+
+
+def multiindex_to_tuple(index: pd.MultiIndex) -> tuple:
+    return list(index.names), list(index.to_list())
+
+
+def tuple_to_multiindex(d: tuple) -> pd.MultiIndex:
+    names, list_ = d
+    return pd.MultiIndex.from_tuples(list_, names=names)
+
+
+class ArrayPacker:
+    """
+    A class to handle converting xarray datasets to and from numpy arrays.
+
+    Used for ML training/prediction.
+    """
+
+    def __init__(
+        self,
+        sample_dim_name,
+        pack_names: Iterable[Hashable],
+        config: Optional[PackerConfig] = None,
+    ):
+        """Initialize the ArrayPacker.
+
+        Args:
+            sample_dim_name: dimension name to treat as the sample dimension
+            pack_names: variable pack_names to pack
+            config: optional PackerConfig for configuration of dataset packing
+        """
+        self._pack_names: List[str] = list(str(s) for s in pack_names)
+        self._n_features: Dict[str, int] = {}
+        self._sample_dim_name = sample_dim_name
+        self._feature_index: Optional[pd.MultiIndex] = None
+        self._config = config
+
+    @property
+    def pack_names(self) -> List[str]:
+        """variable pack_names being packed"""
+        return self._pack_names
+
+    @property
+    def sample_dim_name(self) -> str:
+        """name of sample dimension"""
+        return self._sample_dim_name
+
+    @property
+    def feature_counts(self) -> dict:
+        return self._n_features.copy()
+
+    @property
+    def _total_features(self):
+        return sum(self._n_features[name] for name in self._pack_names)
+
+    def to_array(self, dataset: xr.Dataset) -> np.ndarray:
+        """Convert dataset into a 2D array with [sample, feature] dimensions.
+
+        Variable names inserted into the array are passed on initialization of this
+        object. Each of those variables in the dataset must have the sample
+        dimension name indicated when this object was initialized, and at most one
+        more dimension, considered the feature dimension.
+
+        The first time this is called, the length of the feature dimension for each
+        variable is stored, and can be retrieved on `packer.feature_counts`.
+
+        Args:
+            dataset: dataset containing variables in self.pack_names to pack
+
+        Returns:
+            array: 2D [sample, feature] array with data from the dataset
+        """
+        array, feature_index = pack(
+            dataset[self._pack_names], [self._sample_dim_name], config=self._config
+        )
+        self._n_features = count_features(feature_index)
+        self._feature_index = feature_index
+        return array
+
+    def to_dataset(self, array: np.ndarray) -> xr.Dataset:
+        """Restore a dataset from a 2D [sample, feature] array.
+
+        Can only be called after `to_array` is called.
+
+        Args:
+            array: 2D [sample, feature] array
+
+        Returns:
+            dataset: xarray dataset with data from the given array
+        """
+        if len(array.shape) > 2:
+            raise NotImplementedError("can only restore 2D arrays to datasets")
+        if len(self._n_features) == 0 or self._feature_index is None:
+            raise RuntimeError(
+                "must pack at least once before unpacking, "
+                "so dimension lengths are known"
+            )
+        return unpack(array, [self._sample_dim_name], feature_index=self._feature_index)
+
+    def dump(self, f: TextIO):
+        return yaml.safe_dump(
+            {
+                "pack_names": self._pack_names,
+                "sample_dim_name": self._sample_dim_name,
+                "feature_index": multiindex_to_tuple(self._feature_index),
+                "packer_config": dataclasses.asdict(self._config)
+                if self._config is not None
+                else {},
+            },
+            f,
+        )
+
+    @classmethod
+    def load(cls, f: TextIO):
+        data = yaml.safe_load(f.read())
+        packer_config = dacite.from_dict(PackerConfig, data.get("packer_config", {}))
+        packer = cls(data["sample_dim_name"], data["pack_names"], packer_config)
+        packer._feature_index = tuple_to_multiindex(data["feature_index"])
+        packer._n_features = count_features(packer._feature_index)
+        return packer
+
+
+def get_unpack_layer(array_packer: ArrayPacker, feature_dim: int):
+    return Unpack(
+        pack_names=array_packer.pack_names,
+        n_features=array_packer.feature_counts,
+        feature_dim=feature_dim,
+    )
+
+
+def get_input_vector(
+    packer: ArrayPacker, n_window: Optional[int] = None, series: bool = True,
+):
+    """
+    Given a packer, return a list of input layers with one layer
+    for each input used by the packer, and a list of output tensors which are
+    the result of packing those input layers.
+
+    Args:
+        packer
+        n_window: required if series is True, number of timesteps in a sample
+        series: if True, returned inputs have shape [n_window, n_features], otherwise
+            they are 1D [n_features] arrays
+    """
+    features = [packer.feature_counts[name] for name in packer.pack_names]
+    if series:
+        if n_window is None:
+            raise TypeError("n_window is required if series is True")
+        input_layers = [
+            tf.keras.layers.Input(shape=[n_window, n_features])
+            for n_features in features
+        ]
+    else:
+        input_layers = [
+            tf.keras.layers.Input(shape=[n_features]) for n_features in features
+        ]
+    packed = tf.keras.layers.Concatenate()(input_layers)
+    return input_layers, packed
 
 
 class _BPTTTrainer:
