+++ external/fv3fit/fv3fit/keras/_models/shared/__init__.py
@@ -10,7 +10,7 @@ from .pure_keras import PureKerasModel
 from .training_loop import TrainingLoopConfig, EpochResult
 from .callbacks import TrainingLoopLossHistory
 from .loss import LossConfig
-from .utils import get_input_vector, standard_denormalize, standard_normalize
+from .utils import standard_denormalize, standard_normalize
 from .halos import append_halos
 from .clip import ClipConfig
 from .output_limit import OutputLimitConfig
