+++ external/fv3fit/tests/test_scalers.py
@@ -1,18 +1,13 @@
 import numpy as np
 import pytest
 import tempfile
-from typing import Mapping, Sequence
-import xarray as xr
 
 from fv3fit._shared.scaler import (
     dumps,
     loads,
     StandardScaler,
     ManualScaler,
-    _create_scaling_array,
-    get_mass_scaler,
 )
-from fv3fit._shared.packer import ArrayPacker
 
 SAMPLE_DIM = "sample"
 FEATURE_DIM = "z"
@@ -79,78 +74,6 @@ def test_normalize_then_denormalize_on_reloaded_scaler(n_samples, n_features):
     np.testing.assert_almost_equal(result, X)
 
 
-def _dataset_from_mapping(mapping: Mapping[str, Sequence[float]]):
-    data_vars = {}
-    for var, values in mapping.items():
-        array = np.array(values)
-        if array.shape == (1,):
-            dims = [SAMPLE_DIM]
-        else:
-            dims = [SAMPLE_DIM, FEATURE_DIM]
-            array = np.reshape(array, (1, -1))
-        data_vars[var] = (dims, array)
-    return xr.Dataset(data_vars)
-
-
-@pytest.mark.parametrize(
-    "output_values, delp_weights, \
-        variable_scale_factors, sqrt_weights, expected_weights",
-    [
-        pytest.param(
-            {"y0": [[0.0, 1]], "y1": [[2.0, 3]]},
-            np.array([1.0, 2.0]),
-            {"y0": 100},
-            False,
-            [[100 / 3, 200 / 3, 1.0 / 3, 2.0 / 3]],
-            id="all vertical features, with scale factor",
-        ),
-        pytest.param(
-            {"y0": [[0.0, 1]], "y1": [2]},
-            np.array([1.0, 2.0]),
-            None,
-            False,
-            [[1.0 / 3.0, 2.0 / 3.0, 1.0]],
-            id="one scalar feature",
-        ),
-        pytest.param(
-            {"y0": [0.0], "y1": [2]},
-            np.array([1.0, 2.0]),
-            None,
-            False,
-            [[1.0, 1.0]],
-            id="all scalar features",
-        ),
-        pytest.param(
-            {"y0": [[0.0, 6.0]], "y1": [[4.0, 18.0]]},
-            np.array([4.0, 9.0]),
-            {"y0": 100},
-            True,
-            [
-                [
-                    20 / np.sqrt(13),
-                    30 / np.sqrt(13),
-                    2.0 / np.sqrt(13),
-                    3.0 / np.sqrt(13),
-                ]
-            ],
-            id="sqrt delp, but not scale factors",
-        ),
-    ],
-)
-def test__create_scaling_array(
-    output_values, delp_weights, variable_scale_factors, sqrt_weights, expected_weights
-):
-    ds = _dataset_from_mapping(output_values)
-    packer = ArrayPacker(
-        sample_dim_name=SAMPLE_DIM, pack_names=sorted(list(ds.data_vars))
-    )
-    _ = packer.to_array(ds)
-    weights = _create_scaling_array(
-        packer, delp_weights, variable_scale_factors, sqrt_weights,
-    )
-    np.testing.assert_almost_equal(weights, expected_weights)
-
-
 def test_weight_scaler_denormalize():
     y = np.array([[0.0, 1.0, 2.0, 3.0]])
     weights = np.array([[1.0, 100.0, 1.0, 2.0]])
@@ -179,22 +102,6 @@ def test_weight_scaler_normalize_then_denormalize_on_reloaded_scaler():
             scaler = scaler.load(f_read)
     result = scaler.denormalize(result)
     np.testing.assert_almost_equal(result, y)
-
-
-def test_get_mass_scaler():
-    ds = _dataset_from_mapping({"y0": [[0.0, 3.0]], "y1": [[2.0, 6.0]]})
-    packer = ArrayPacker(
-        sample_dim_name=SAMPLE_DIM, pack_names=sorted(list(ds.data_vars))
-    )
-    y = packer.to_array(ds)
-    delp = np.array([1.0, 0.25])
-    scale_factors = {"y0": 100}
-    scaler = get_mass_scaler(packer, delp, scale_factors, True,)
-    expected_normalized = [
-        [0.0, 15 / np.sqrt(1.25), 2 / np.sqrt(1.25), 3 / np.sqrt(1.25)]
-    ]
-    normalized = scaler.normalize(y)
-    np.testing.assert_almost_equal(normalized, expected_normalized)
 
 
 standard_scaler = StandardScaler()
